% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDDConf-package.R
\docType{package}
\name{IDDConf-package}
\alias{IDDConf}
\alias{IDDConf-package}
\title{IDDConf: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SHVickers94/IDDConf}
  \item Report bugs at \url{https://github.com/SHVickers94/IDDConf/issues}
}

}
\author{
\strong{Maintainer}: Stephen Vickers \email{svickers@rvc.ac.uk} (\href{https://orcid.org/0000-0002-7809-8904}{ORCID}) [copyright holder]

}
\keyword{internal}
